Refactored direction to take inner radius as argument
