+++ Sources/LCARSDisplayKit/Shapes/CoreGraphics+Translation.swift
@@ -3,6 +3,7 @@ import GraphPoint
 import CoreGraphics
 
 extension CGMutablePath {
+    @available(*, deprecated)
     func addArc(center: CGPoint, radius: Radius, startAngle: Degree, endAngle: Degree, clockwise: Bool) {
         addArc(
             center: center,
@@ -25,6 +26,14 @@ extension CGMutablePath {
     
     func addLine(to point: CartesianPoint) {
         addLine(to: CGPoint(point))
+    }
+    
+    func move(to point: CartesianPoint) {
+        move(to: CGPoint(point))
+    }
+    
+    func addRect(_ rect: CartesianFrame) {
+        addRect(CGRect(rect))
     }
 }
